+++ src/types.ts
@@ -1,22 +1,5 @@
 import { UserQuery } from './utils/__generated__/graphql-request';
 
-export type SocialProvider =
-  | 'apple'
-  | 'azuread'
-  | 'bitbucket'
-  | 'discord'
-  | 'facebook'
-  | 'github'
-  | 'gitlab'
-  | 'google'
-  | 'linkedin'
-  | 'spotify'
-  | 'strava'
-  | 'twitch'
-  | 'twitter'
-  | 'windowslive'
-  | 'workos';
-
 export type ClaimValueType =
   | string
   | string[]